Consolidate some files and flatten hierarchy
