--- src/core/mod.rs
@@ -1,10 +1,0 @@
-mod editor;
-mod line;
-mod selection;
-
-// CodeLines fields are private and need to be
-// explicitly re-exported.
-pub use editor::CodeLinesStoreFields;
-pub use editor::Editor;
-pub use line::EditLine;
-pub use selection::SetSelection;
