+++ src/lib.rs
@@ -1,3 +1,4 @@
-pub mod core;
+pub mod editor;
+pub mod line;
 pub mod utils;
 pub mod view;
