--- src/view/mod.rs
@@ -1,2 +1,0 @@
-pub mod component;
-pub mod dom;
